requirements.txt
flask
gunicorn
google-generativeai
sqlalchemy
pymysql
pandas
python-dotenv
cryptography
pymssql